# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      consumes:
      - application/json
      description: 可按社区按时间或分数排序查询帖子列表接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 升级版帖子列表接口
      tags:
      - HomeWebPage
  /user/createUser:
    get:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: repassword
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      tags:
      - 用户模块
  /user/deleteUser:
    get:
      parameters:
      - description: 用户id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      tags:
      - 用户模块
  /user/findUserByNameAndPwd:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - HomeWebPage
  /user/getlist:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - HomeWebPage
  /user/updateUser:
    post:
      parameters:
      - description: 用户id
        in: formData
        name: id
        type: string
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      tags:
      - 用户模块
swagger: "2.0"